Implement stock take for non-lot managed trade items
