+++ openlmis-stock/src/main/res/layout/stock_take_trade_item_row.xml
@@ -55,48 +55,6 @@
 
     </LinearLayout>
 
-    <RelativeLayout
-        android:id="@+id/completed_stock_take"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="8dp"
-        android:visibility="gone">
-
-        <ImageView
-            android:id="@+id/check"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentStart="true"
-            android:layout_margin="10dp"
-            android:background="@drawable/ic_check_circle" />
-
-        <TextView
-            android:id="@+id/completed_trade_item"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10dp"
-            android:layout_toEndOf="@id/check"
-            android:textColor="@color/text_black"
-            android:textSize="18sp"
-            android:textStyle="bold" />
-
-        <TextView
-            android:id="@+id/adjustment"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/completed_trade_item"
-            android:layout_toEndOf="@id/check"
-            android:textColor="@color/stock_list_light_text"
-            android:textSize="12sp" />
-
-
-        <ImageView
-            android:id="@+id/edit_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:layout_margin="10dp"
-            android:src="@drawable/ic_stock_edit" />
-    </RelativeLayout>
+    <include layout="@layout/stock_take_trade_item_completed" />
 
 </android.support.v7.widget.CardView>
